requirements.txt
pandas
uvicorn
python-dotenv
langchain_groq
langchain==0.1.0
langchain-experimental==0.0.49
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0
tabulate==0.9.0
fastapi
scikit-learn
python-multipart